- Moved more-so “utility” functions into `time.ts` and `string.ts`. - Moved CMU Eats API url to `App.tsx`
